chore(v3/label_inventories): rename file
fix(v3/urls): fix bad import

chore: reorganize imports
