+++ seed/views/v3/label_inventories.py
